+++ src/Test/PocketBaseClient.DemoTest/Models/TestForRelated.cs
@@ -32,28 +32,28 @@ namespace PocketBaseClient.DemoTest.Models
         #endregion Collection
 
         #region Field Properties
-        private int? _NumberUnique = null;
+        private float? _NumberUnique = null;
         /// <summary> Maps to 'number_unique' field in PocketBase </summary>
         [JsonPropertyName("number_unique")]
         [PocketBaseField(id: "s10g39sb", name: "number_unique", required: false, system: false, unique: true, type: "number")]
         [Display(Name = "Number unique")]
-        public int? NumberUnique { get => Get(() => _NumberUnique); set => Set(value, ref _NumberUnique); }
+        public float? NumberUnique { get => Get(() => _NumberUnique); set => Set(value, ref _NumberUnique); }
 
-        private int? _NumberNonempty = null;
+        private float? _NumberNonempty = null;
         /// <summary> Maps to 'number_nonempty' field in PocketBase </summary>
         [JsonPropertyName("number_nonempty")]
         [PocketBaseField(id: "iy8rrkm2", name: "number_nonempty", required: true, system: false, unique: false, type: "number")]
         [Display(Name = "Number nonempty")]
         [Required(ErrorMessage = @"NumberNonempty is required")]
-        public int? NumberNonempty { get => Get(() => _NumberNonempty); set => Set(value, ref _NumberNonempty); }
+        public float? NumberNonempty { get => Get(() => _NumberNonempty); set => Set(value, ref _NumberNonempty); }
 
-        private int? _NumberNonemptyUnique = null;
+        private float? _NumberNonemptyUnique = null;
         /// <summary> Maps to 'number_nonempty_unique' field in PocketBase </summary>
         [JsonPropertyName("number_nonempty_unique")]
         [PocketBaseField(id: "mmzxqln4", name: "number_nonempty_unique", required: true, system: false, unique: true, type: "number")]
         [Display(Name = "Number nonempty unique")]
         [Required(ErrorMessage = @"NumberNonemptyUnique is required")]
-        public int? NumberNonemptyUnique { get => Get(() => _NumberNonemptyUnique); set => Set(value, ref _NumberNonemptyUnique); }
+        public float? NumberNonemptyUnique { get => Get(() => _NumberNonemptyUnique); set => Set(value, ref _NumberNonemptyUnique); }
 
         #endregion Field Properties
